frontend: impl useHttp hook with auth header
