+++ frontend/src/components/TaskForm.tsx
@@ -5,8 +5,7 @@ import { useForm } from "react-hook-form";
 
 import TaskContext from "../contexts/task-context";
 import { Task } from "../types/task";
-import client from "../utils/api";
-import AuthContext from "../contexts/auth-context";
+import useHttp from "../hooks/useHttp";
 
 type FormValues = {
   name: string;
@@ -17,19 +16,13 @@ const TaskForm: React.FC = () => {
 
   const taskCtx = useContext(TaskContext);
 
-  const authCtx = useContext(AuthContext);
-
-  const authHead = {
-    headers: {
-      Authorization: `Bearer ${authCtx.token}`,
-    },
-  };
+  const { client } = useHttp();
 
   // Handler for Add Task button.
   const addTaskHandler = (task: Task) => {
     // send POST request and append the response data to array `tasks`.
     // The response comes with an ID generated by the backend DB.
-    client.post("/tasks", task, authHead).then((response) => {
+    client.post("/tasks", task).then((response) => {
       if (response.data.error !== undefined && response.data.error !== "") {
         alert(`Error from server: ${response.data.error}`);
         return;
